Use terraform plan to determine whether the workspace is up-to-date
Previously we assumed the workspace was not up-to-date during the observe phase
and went straight to invoking terraform apply. The rationale was we might as
well go straight to trying to apply, because terraform will just abort without
making any changes if none are needed. The downside of this approach is that the
resource would act as if it were constantly out-of-date, emitting a spurious
update event on every reconcile.

Signed-off-by: Nic Cope <negz@rk0n.org>